refactor(agent-templates): split 1395-line file into focused modules (#230)
Replaces the single packages/web/src/lib/agent-templates.ts with a directory
of small modules organised by concern:

- types.ts                  — type definitions
- odoo-factory.ts           — createOdooTemplate + ODOO_* prompt constants
- generate-agents-md.ts     — generateAgentsMd rendering
- data/knowledge-base.ts    — knowledge-base, custom
- data/document-agents.ts   — contract-analyzer, resume-screener, etc.
- data/odoo-agents.ts       — all 17 odoo-* templates
- data/email-agents.ts      — email-assistant variants
- registry.ts               — AGENT_TEMPLATES, getTemplate, getTemplateList,
                              pickSuggestedName
- index.ts                  — re-exports the public surface unchanged

No behaviour change: every callable and value is byte-identical, the public
import surface (@/lib/agent-templates) is preserved, and the existing test
suite passes without modification.
